Protect APIs with Authentication

+++ backend/routes/todos.js
@@ -6,19 +6,19 @@ const auth = require("../middleware/auth");
 //Create instance of Router
 const router = express.Router();
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
   try {
     const todos = await Todo.find().sort({ date: -1 });
     //console.log(req.user);
-
-    res.send(todos);
+    const filteredTodos = todos.filter((todo) => todo.uid === req.user._id);
+    res.send(filteredTodos);
   } catch (error) {
     res.status(500).send(error.message);
     console.log(error.message);
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   //Add Joi validation
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
@@ -65,6 +65,7 @@ router.post("/", async (req, res) => {
   }
 });
 
+/*
 router.delete("/", async (req, res) => {
   //deleteOne() If there are multiple that satisfy the filter, the first
   //one that is found will be deleted
@@ -77,15 +78,19 @@ router.delete("/", async (req, res) => {
     res.status(500).send(error.message);
     console.log(error.message);
   }
-});
+});*/
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     //Check to see if a todo with the given id exists
     const todo = await Todo.findById(req.params.id);
 
     //Send Error 404 if there is no To do
     if (!todo) return res.status(404).send("Todo Not Found...");
+
+    //Send Error 401 if user isnt creator of the To do
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo deletion failed. Not authorized...");
 
     //findByIdAndDelete()
     const deletedTodo = await Todo.findByIdAndDelete(req.params.id);
@@ -97,7 +102,7 @@ router.delete("/:id", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   //Add Joi validation
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
@@ -120,6 +125,10 @@ router.put("/:id", async (req, res) => {
     //Send Error 404 if there is no To do
     if (!todo) return res.status(404).send("Todo Not Found...");
 
+    //Send Error 401 if user isnt creator of the To do
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo update failed. Not authorized...");
+
     //Get update parameters from API call
     const { name, author, isComplete, date, uid } = req.body;
 
@@ -137,17 +146,27 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     //Check to see if a todo with the given id exists
     const todo = await Todo.findById(req.params.id);
     //Send Error 404 if there is no To do
     if (!todo) return res.status(404).send("Todo Not Found...");
 
-    const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, {
-      //Set isComplete to the opposite value
-      isComplete: !todo.isComplete,
-    }, {new: true});
+    //Send Error 401 if user isnt creator of the To do
+    if (todo.uid !== req.user._id)
+      return res
+        .status(401)
+        .send("Todo check/uncheck failed. Not authorized...");
+
+    const updatedTodo = await Todo.findByIdAndUpdate(
+      req.params.id,
+      {
+        //Set isComplete to the opposite value
+        isComplete: !todo.isComplete,
+      },
+      { new: true }
+    );
 
     res.send(updatedTodo);
   } catch (error) {
